[EC-115] Moved validators to its own package
