+++ src/main/scala/io/iohk/ethereum/ledger/Ledger.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.ledger
 
 import io.iohk.ethereum.db.storage.NodeStorage
 import io.iohk.ethereum.domain._
-import io.iohk.ethereum.network.p2p.validators.SignedTransactionValidator
+import io.iohk.ethereum.validators.SignedTransactionValidator
 import io.iohk.ethereum.utils.{Config, Logger}
 import io.iohk.ethereum.vm.{GasFee, _}
