+++ src/main/scala/io/iohk/ethereum/validators/MptListValidator.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.network.p2p.validators
+package io.iohk.ethereum.validators
 
 import io.iohk.ethereum.rlp.RLPImplicits._
 import io.iohk.ethereum.crypto.kec256
